Add new appLoadingSlice, remove appAction, reducer
